requirements.txt
asgiref==3.7.2
astroid==3.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.7
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
filelock==3.13.1
gunicorn==22.0.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
packaging==24.0
panda==0.3.1
pillow==10.3.0
platformdirs==4.1.0
prettytable==3.9.0
psycopg2-binary==2.9.9
pygame==2.5.2
PyJWT==2.8.0
pylint==3.0.3
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
whitenoise==6.6.0
